Add script to test tasks restarts. add imagenet gcs script for TPUs. Add ability to override the queue.
- Add ability to import .dva_model_export.zip via perform_import
- Training sets for Cloud TPUs
- YFCC prep scripts to VDN.

Next:

- Implement pre-condition checks, failed to launch error messages for
  processes. Rollback created models on failure to launch. Add
  consistency checks regarding video_selector being in the
  arguments to jsonschema.

+++ server/dvaapp/task_shared.py
@@ -32,9 +32,10 @@ def restart_task(dt, exception_traceback):
             return None
         else:
             logging.info("Restarting {}".format(dt.pk))
-            for model_name in settings.RESTARTABLE_TASKS[dt.operation]['delete_models']:
-                m = apps.get_model(app_label='dvaapp', model_name=model_name)
-                m.objects.filter(event_id=dt.pk).delete()
+            if 'delete_models' in settings.RESTARTABLE_TASKS[dt.operation]:
+                for model_name in settings.RESTARTABLE_TASKS[dt.operation]:
+                    m = apps.get_model(app_label='dvaapp', model_name=model_name)
+                    m.objects.filter(event_id=dt.pk).delete()
             new_dt = TEvent.objects.create(parent_process=dt.parent_process,
                                            task_group_id=dt.task_group_id,
                                            parent=dt.parent,
